healthproj/requirements.txt
pytz==2020.1
django-widget-tweaks==1.4.8